+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- `BoxedWithLength` wrapper type which attaches id and size_hint of a `T: Identifiable + MtProtoSized` type.
+- `WithSize` wrapper type which attaches size hint of a `T: MtProtoSized` type.
+- `BoxedWithSize` wrapper type which attaches id and size hint of a `T: Identifiable + MtProtoSized` type.
 - `MtProtoSized` implementation for arrays up to length 32 and `Box<T>` where `T: MtProtoSized`.
 - Size hints for `de::SeqAccess` and `de::MapAccess`.
 - `UnsizedByteBuf` helper type and tests for it.